refactor: move around tests for rank operator
